requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-mail==1.4.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
rfc3986==1.5.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
